# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Biblioteca con PostgreSQL",
    "description": "Sistema de gestión de biblioteca con autenticación JWT y PostgreSQL",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Token JWT en formato: Bearer <token>"
    }
  },
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/registro": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Registrar nuevo usuario",
        "description": "Crea una nueva cuenta de usuario",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Datos del usuario",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "usuario@ejemplo.com"
                },
                "password": {
                  "type": "string",
                  "example": "contraseña123"
                },
                "nombre": {
                  "type": "string",
                  "example": "Juan Pérez"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Iniciar sesión",
        "description": "Autentica un usuario y devuelve un token JWT",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Credenciales de usuario",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "usuario@ejemplo.com"
                },
                "password": {
                  "type": "string",
                  "example": "contraseña123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/perfil": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Obtener perfil del usuario actual",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/": {
      "get": {
        "tags": [
          "Votaciones"
        ],
        "summary": "Listar todas las votaciones",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Votaciones"
        ],
        "summary": "Crear nueva votación",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Datos de la votación",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "titulo": {
                  "type": "string",
                  "example": "Mejor Libro del Año"
                },
                "descripcion": {
                  "type": "string",
                  "example": "Votación para elegir el mejor libro"
                },
                "opciones": {
                  "type": "array",
                  "example": [
                    "Opción 1",
                    "Opción 2",
                    "Opción 3"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "fecha_fin": {
                  "type": "string",
                  "example": "2024-12-31T23:59:59Z"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/{id}/rol": {
      "patch": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Actualizar rol de usuario",
        "description": "Solo accesible para administradores",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del usuario"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Nuevo rol",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "rol": {
                  "type": "string",
                  "example": "admin"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/{id}": {
      "get": {
        "tags": [
          "Votaciones"
        ],
        "summary": "Obtener votación por ID con resultados",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Votaciones"
        ],
        "summary": "Actualizar votación",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "titulo": {
                  "example": "any"
                },
                "descripcion": {
                  "example": "any"
                },
                "fecha_fin": {
                  "example": "any"
                },
                "activa": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Votaciones"
        ],
        "summary": "Eliminar votación",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/{id}/votar": {
      "post": {
        "tags": [
          "Votaciones"
        ],
        "summary": "Votar en una votación",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID de la votación"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Opción de voto",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "opcion": {
                  "type": "string",
                  "example": "Opción 1"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
    "Usuario": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "uuid"
        },
        "email": {
          "type": "string",
          "example": "usuario@ejemplo.com"
        },
        "nombre": {
          "type": "string",
          "example": "Juan Pérez"
        },
        "rol": {
          "type": "string",
          "example": "usuario"
        },
        "created_at": {
          "type": "string",
          "example": "2024-01-01T00:00:00Z"
        }
      }
    },
    "Evento": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "uuid"
        },
        "titulo": {
          "type": "string",
          "example": "Conferencia de Literatura"
        },
        "descripcion": {
          "type": "string",
          "example": "Evento sobre literatura contemporánea"
        },
        "fecha": {
          "type": "string",
          "example": "2024-12-31T18:00:00Z"
        },
        "imagen_url": {
          "type": "string",
          "example": "http://localhost:3000/uploads/eventos/imagen.jpg"
        },
        "created_by": {
          "type": "string",
          "example": "uuid"
        },
        "created_at": {
          "type": "string",
          "example": "2024-01-01T00:00:00Z"
        }
      }
    },
    "Funcion": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "uuid"
        },
        "titulo": {
          "type": "string",
          "example": "Obra de Teatro"
        },
        "descripcion": {
          "type": "string",
          "example": "Presentación teatral"
        },
        "fecha": {
          "type": "string",
          "example": "2024-12-31T20:00:00Z"
        },
        "imagen_url": {
          "type": "string",
          "example": "http://localhost:3000/uploads/funciones/imagen.jpg"
        },
        "created_by": {
          "type": "string",
          "example": "uuid"
        },
        "created_at": {
          "type": "string",
          "example": "2024-01-01T00:00:00Z"
        }
      }
    },
    "Votacion": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "uuid"
        },
        "titulo": {
          "type": "string",
          "example": "Mejor Libro del Año"
        },
        "descripcion": {
          "type": "string",
          "example": "Votación para elegir el mejor libro"
        },
        "opciones": {
          "type": "array",
          "example": [
            "Opción 1",
            "Opción 2",
            "Opción 3"
          ],
          "items": {
            "type": "string"
          }
        },
        "fecha_fin": {
          "type": "string",
          "example": "2024-12-31T23:59:59Z"
        },
        "activa": {
          "type": "boolean",
          "example": true
        },
        "created_by": {
          "type": "string",
          "example": "uuid"
        },
        "created_at": {
          "type": "string",
          "example": "2024-01-01T00:00:00Z"
        }
      }
    }
  }
}